feat(game-card): add game card component
readability
